app/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
neo4j==5.16.0
psycopg2-binary==2.9.9
pandas==2.1.4
python-dotenv==1.0.0